+++ packages/terra-worklist-data-grid/tests/jest/ColumnHeaderCell.test.jsx
@@ -5,7 +5,7 @@ import IconUp from 'terra-icon/lib/icon/IconUp';
 import IconDown from 'terra-icon/lib/icon/IconDown';
 import IconError from 'terra-icon/lib/icon/IconError';
 import { shallowWithIntl } from 'terra-enzyme-intl';
-import ColumnHeaderCell from '../../src/ColumnHeaderCell';
+import ColumnHeaderCell from '../../src/subcomponents/ColumnHeaderCell';
 
 describe('ColumnHeaderCell', () => {
   it('render default column header cell', () => {
@@ -16,7 +16,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
       />,
@@ -27,7 +27,7 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
@@ -47,7 +47,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
       />,
@@ -58,7 +58,7 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
@@ -79,7 +79,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
       />,
@@ -90,7 +90,7 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('descending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
@@ -111,7 +111,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
       />,
@@ -122,7 +122,7 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBeUndefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
@@ -144,7 +144,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
       />,
@@ -155,7 +155,7 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
@@ -180,7 +180,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
         onColumnSelect={mockClick}
@@ -192,9 +192,9 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
-    expect(columnHeader.props().onClick).toBeDefined();
+    expect(columnHeader.props().onMouseDown).toBeDefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
@@ -219,7 +219,7 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
+        {...column}
         width="100px"
         headerHeight="150px"
         onColumnSelect={onColumnSelect}
@@ -231,9 +231,9 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
-    expect(columnHeader.props().onClick).toBeUndefined();
+    expect(columnHeader.props().onMouseDown).toBeUndefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');