+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,10 +5,13 @@
 * Changed
   * Updated example for `terra-menu`.
   * Updated tests for `terra-worklist-data-grid` to account for row height and column width being honored.
+  * Updated the path name for the examples in doc site for `terra-embedded-content-consumer` to resolve missing files.
   * Updated examples for `terra-tabs` with conflicting ids.
 
 * Added
   * Added alpha release disclaimer for `terra-worklist-data-grid`.
+  * Added documentation and tests for additional keyboard navigation in `terra-worklist-data-grid`.
+  * Added Cell Selection example for `terra-worklist-data-grid`.
 
 ## 1.24.0 - (July 4, 2023)